+++ routes/activities.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const { post } = require('../controllers/activities')
 const { schemaValidator } = require('../middlewares/validator')
-const { activity } = require('../schemas/activity')
+const { activity } = require('../schemas/activities')
 
 const router = express.Router()
