+++ schemas/activities.js
